pyproject.toml
[project]
name = "hr-agent-poc"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "semantic-kernel",
  "azure-identity",
  "azure-search-documents",
  "azure-monitor-opentelemetry",
  "azure-cosmos",
  "python-dotenv"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
